Rename test modules with "test_" prefixes.
