+++ sandpiper/tests/__init__.py
@@ -1,3 +1,3 @@
-from .bios import *
-from .conversion import *
-from .database import *
+from .test_bios import *
+from .test_conversion import *
+from .test_database import *
